Fix type conversion in warpBitmapWithCorners
Corrected the conversion of corner coordinates to ensure they are passed as Double values to the OpenCV Point constructor. Also updated release notes for version 1.0.6 with new features and changes.

+++ app/src/main/java/app/papra/mobile/ui/ScanUtils.kt
@@ -67,7 +67,7 @@ fun warpBitmapWithCorners(bitmap: Bitmap, corners: List<PointF>): Bitmap {
     val src = Mat()
     Utils.bitmapToMat(bitmap, src)
     val srcPoints = corners.map {
-        Point(it.x * bitmap.width, it.y * bitmap.height)
+        Point((it.x * bitmap.width).toDouble(), (it.y * bitmap.height).toDouble())
     }
     val ordered = orderPoints(srcPoints.toTypedArray())
     val (tl, tr, br, bl) = ordered
